Fix tags services & controllers

+++ src/app/tags/tags.service.ts
@@ -4,26 +4,27 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Tag } from './entities/tag.entity';
 import { CreateTagDto } from './dto/create-tag.dto';
 import { UpdateTagDto } from './dto/update-tag.dto';
+import { DbConnectionName } from '../db/db.config';
 
 @Injectable()
 export class TagsService {
   constructor(
-    @InjectRepository(Tag, 'reader')
+    @InjectRepository(Tag, DbConnectionName.READER)
     private readonly readerRepository: Repository<Tag>,
-    @InjectRepository(Tag, 'editor')
+    @InjectRepository(Tag, DbConnectionName.EDITOR)
     private readonly editorRepository: Repository<Tag>,
   ) {}
 
-  create(tagName: CreateTagDto): Promise<Tag> {
-    const tag = this.editorRepository.create({ tag: tagName.tag });
+  create(newTag: CreateTagDto): Promise<Tag> {
+    const tag = this.editorRepository.create({ tag: newTag.name });
     return this.editorRepository.save(tag);
   }
 
-  async findAll(): Promise<Tag[]> {
+  findAll(): Promise<Tag[]> {
     return this.readerRepository.find();
   }
 
-  findOne(id: number): Promise<Tag | null> {
+  findById(id: number): Promise<Tag | null> {
     return this.readerRepository.findOneBy({ id });
   }
 
@@ -33,11 +34,17 @@ export class TagsService {
 
   async update(id: number, updatedTag: UpdateTagDto): Promise<Tag | null> {
     return this.editorRepository
-      .update({ id }, updatedTag)
-      .then(() => this.findOne(id));
+      .update({ id }, { tag: updatedTag.name })
+      .then(() => this.findById(id));
   }
 
-  async remove(id: number): Promise<void> {
-    await this.editorRepository.delete({ id });
+  async remove(id: number): Promise<Tag | null> {
+    const tag = await this.findById(id);
+    if (!tag) {
+      return null;
+    } else {
+      await this.editorRepository.delete({ id });
+      return tag;
+    }
   }
 }
